requirements.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.15
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==6.2.1
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
requests==2.22.0
uritemplate==3.0.1
urllib3==1.25.7
